+++ packages/system/vyuh_feature_auth/lib/ui/auth_form_builder.dart
@@ -1,19 +1,21 @@
-import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:material_ui/material_ui.dart';
+import 'package:reactive_forms/reactive_forms.dart';
 import 'package:vyuh_feature_auth/ui/auth_state_widget.dart';
 import 'package:vyuh_feature_auth/ui/form_fields.dart';
 
 final class AuthFormBuilder extends StatelessWidget {
+  final FormGroup Function() form;
   final Widget Function(BuildContext, AuthFlowScope, VoidCallback submit) child;
   final Widget Function(BuildContext, AuthFlowScope)? footer;
   final String actionTitle;
-  final Future<void> Function(FormBuilderState) authAction;
+  final Future<void> Function(FormGroup) authAction;
   final AuthState endAuthState;
   final bool showError;
   final ErrorBuilder? errorBuilder;
 
   const AuthFormBuilder({
     super.key,
+    required this.form,
     required this.child,
     required this.actionTitle,
     this.footer,
@@ -27,14 +29,15 @@ final class AuthFormBuilder extends StatelessWidget {
   Widget build(BuildContext context) {
     return AuthFlow(
       builder: (context, scope) {
-        return FormBuilder(
-          child: Column(
+        return ReactiveFormBuilder(
+          form: form,
+          builder: (context, formGroup, _) => Column(
             mainAxisAlignment: MainAxisAlignment.center,
             crossAxisAlignment: CrossAxisAlignment.stretch,
             children: [
               Builder(
                 builder: (context) {
-                  submit() => _submit(context, scope);
+                  submit() => _submit(formGroup, scope);
                   return child(context, scope, submit);
                 },
               ),
@@ -42,7 +45,7 @@ final class AuthFormBuilder extends StatelessWidget {
               AuthActionButton(
                 scope: scope,
                 title: actionTitle,
-                onPressed: (context) => _submit(context, scope),
+                onPressed: (_) => _submit(formGroup, scope),
                 showError: showError,
                 errorBuilder: errorBuilder,
               ),
@@ -54,15 +57,12 @@ final class AuthFormBuilder extends StatelessWidget {
     );
   }
 
-  void _submit(BuildContext context, AuthFlowScope scope) {
+  void _submit(FormGroup formGroup, AuthFlowScope scope) {
     FocusManager.instance.primaryFocus?.unfocus();
-    final state = FormBuilder.of(context);
-    if (state == null) {
-      return;
-    }
+    formGroup.markAllAsTouched();
 
-    if (state.saveAndValidate()) {
-      scope.runAuthAction(() => authAction(state), endState: endAuthState);
+    if (formGroup.valid) {
+      scope.runAuthAction(() => authAction(formGroup), endState: endAuthState);
     }
   }
 }